// management/registry-ui/src/generated/client/models/model_Announcement.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type model_Announcement = {
    content?: string;
    created_at?: string;
    id?: number;
    level?: string;
    pinned?: boolean;
    published?: boolean;
    title?: string;
    updated_at?: string;
};